# openapi.yaml
openapi: 3.0.3
info:
  title: Document Analyzer RAG API
  version: 2.0.0
  description: Session-based PDF ingestion and RAG chat backend.
servers:
  - url: http://localhost:4000
components:
  schemas:
    ErrorObject:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ErrorObject'
    SuccessEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: object
    JobState:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        stage:
          type: string
          enum: [uploading, parsing, chunking, embedding, retrieving, generating]
        queuePosition:
          type: integer
          minimum: 0
        attempts:
          type: integer
        result:
          type: object
          nullable: true
        error:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/ErrorObject'
          nullable: true
        metrics:
          type: object
          nullable: true
    ChatResponseSection:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    ChatResponseSchema:
      type: object
      nullable: true
      properties:
        format:
          type: string
          enum: [structured_sections, structured_markdown]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ChatResponseSection'
    ChatResponseData:
      type: object
      properties:
        answer:
          type: string
        formattedAnswer:
          type: string
        responseSchema:
          $ref: '#/components/schemas/ChatResponseSchema'
        responseStyle:
          type: string
          enum: [structured, plain]
        sources:
          type: array
          items:
            type: object
            properties:
              pdfId:
                type: integer
              chunkId:
                type: integer
              score:
                type: number
        usedChunksCount:
          type: integer
        sessionTitle:
          type: string
paths:
  /api/v1/health:
    get:
      summary: Health and runtime metrics
      responses:
        '200':
          description: OK
  /api/v1/ping:
    get:
      summary: Ping endpoint
      responses:
        '200':
          description: OK
  /api/v1/sessions:
    get:
      summary: List sessions
      responses:
        '200':
          description: OK
    post:
      summary: Create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
      responses:
        '200':
          description: Created
        '422':
          description: Validation error
  /api/v1/sessions/{sessionId}:
    get:
      summary: Get session details
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
    delete:
      summary: Delete session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
  /api/v1/sessions/{sessionId}/pdfs:
    get:
      summary: List PDFs by session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
    post:
      summary: Upload PDF for session and enqueue indexing
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        '202':
          description: Accepted
  /api/v1/pdfs/{pdfId}:
    get:
      summary: Get PDF metadata
      parameters:
        - in: path
          name: pdfId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
    delete:
      summary: Delete PDF metadata and optionally stored file
      parameters:
        - in: path
          name: pdfId
          required: true
          schema: { type: integer }
        - in: query
          name: removeFile
          schema: { type: boolean }
      responses:
        '200':
          description: Deleted
  /api/v1/sessions/{sessionId}/chat:
    post:
      summary: Run chat query against indexed session documents
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
        - in: query
          name: stream
          schema:
            type: boolean
          description: Enable SSE token streaming
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      text:
                        type: string
                responseStyle:
                  type: string
                  enum: [structured, plain]
                  description: Response formatting style. Defaults to structured.
      responses:
        '200':
          description: Chat answer or SSE stream response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChatResponseData'
            text/event-stream:
              schema:
                type: string
        '202':
          description: Async job accepted
  /api/v1/jobs/{jobId}:
    get:
      summary: Poll async job status
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JobState'
  /api/v1/sessions/{sessionId}/history:
    get:
      summary: Get chat history
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200':
          description: OK
    delete:
      summary: Clear session history
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Cleared
  /api/v1/admin/queue:
    get:
      summary: Queue/metrics overview (non-production)
      responses:
        '200':
          description: OK
  /api/v1/admin/reset:
    post:
      summary: Reset upload directory (non-production)
      responses:
        '200':
          description: OK